6. U paket „model“ dodati zapis (engl. record) pod nazivom „Discount“ te ga dodati kao novi objekt unutar klase „Item“. Taj zapis mora sadržavati podatak „discountAmount“ koji označava iznos popusta (u postocima). Kod unosa podataka za artikle je od korisnika potrebno odabrati i iznos popusta te kod preračunavanja cijene uzeti i popust u obzir.

+++ src/main/java/hr/java/production/model/Bread.java
@@ -26,7 +26,7 @@ public class Bread extends Item implements Edible {
     }
 
     @Override
-    public double calculatePrice(double weight) {
+    public BigDecimal calculatePrice(double weight) {
         return BigDecimal.valueOf(weight).multiply(super.getSellingPrice()) ;
     }
 }
