+++ src/main/java/hr/java/production/model/Edible.java
@@ -1,7 +1,9 @@
 package hr.java.production.model;
 
+import java.math.BigDecimal;
+
 public interface Edible {
     int calculateKilocalories();
 
-    double calculatePrice(double weight);
+    BigDecimal calculatePrice(double weight);
 }
